Change CassandraConfiguration.java in favour of AbstractCassandraConfiguration.
Previously the mapping context never scanned the entities and hence could not create a cache for models that it was mapping from the cassandra tables. The reason was due to manual configuration and that I have missed certain beans and class loaders because they were never mentioned in the documentation. After load testing and profiling the application, the decision has been made to change
the configuration to the more peformant one provided by Spring Boot. Furthermore, I have fine tuned the performance and added compression and overrode certain configuration props.