MessageService.java implement logic for reading the messages.
Implemented custom reading and aggregation algorithm (in case a time bucket does not have enough of data).

Firstly, if the beforeMessageId is present and the provided bucket is valid, then we resolve the messages based on the message id and well known ordering.

However, if the beforeMessageId is not provided we at least expect the bucket to be present and fetch it based on the bucket and possible paging state (scroll state) if provided. However, if none of them are present, then we create a bucket based on the current timestamp and try to find any messages for the user and thread.

Thereafter, if the messages that we have fetched are below the desired page size, we have to apply an iterative algorithm that generates the buckets from the threadId (which is a TIMEUUID) till the resolvedBucket - 1 until it reaches the end. On the other hand, this step is skipped if the bucket we could not aggregate the data from was the 0th bucket, i.e. there simply is no data in the chat.