Fixed the bug with message aggregation algorithm.
The bug was firstly that the paging state of the previous iteration was being used, the problem is that when we call the method we assume that we have exhausted our initial partition and the paging state is soleley for a single partition and same query string. Therefore, applying paging state to the next partitions simply does not make sense. Secondly, the last paging state that was read before we have reached our goal, we would not save it properly and hence the user would not be able to continue paging from there.

Lastly, I was passing paging state from the previous iteration to the next aggregation query, it again does not make sense because we change the bucket and hence we change partition, therefore, paging state does not apply.